Readding trusted applications
Is needed to be able to render trustedApplications in Solid UI (part of https://github.com/solid/solid-ui/issues/121)
