+++ src/registerPanes.js
@@ -7,6 +7,7 @@ module.exports = function registerPanes (register) {
   register(require('./profile/profilePane')) // View someone's public profile - dominates all other panes.
   register(require('./profile/editProfilePane')) // Edit my profile. App. 201900802
 
+  register(require('./trustedApplications/trustedApplicationsPane')) // must be registered before basicPreferences
   register(require('./dashboard/dashboardPane'))
   register(require('./dashboard/basicPreferences')) // 20190702
   register(require('issue-pane'))
