+++ src/trustedApplications/trustedApplications.test.ts
